Add mock action buttons to StackSync activity
Add action buttons to the StackSync activity's action bar and overflow:
search, notifications, and settings. Don't yet implement the
functionality of these actions. Include an icon for the search button
for a range of screen densities.

The notifications button uses a custom view, NotificationsActionView,
to allow a number representing the current number of unread
notifications to be displayed over the button. Fix number to 5 for the
sake of mocking up. Implement the "tooltip" that occurs when long-
pressing on default action buttons.