Add application launcher icon
Add the nice new launcher icon for four different screen
densities: ldpi, mdpi, hdpi, and xhdpi. The icon is based on the
Stack Exchange logo with an arrow shaped hole in the middle. It is
in a 3D perspective and slightly tilted forward with lighting from
above, as the Android design guidelines ask.