Add fragment for listing sites with dummy data
Add a fragment to be displayed under the "Sites" tab of the StackSync
activity. This fragment lists Stack Exchange sites together with their
icons and other details. When a list item is pressed, it will take the
user to the content of that site. A dropdown menu is found beside each
list item to give more information and shortcuts to deeper into the
hierarchy.

The data is currently dummy data and the site icons are left out of
this commit to avoid copyright issues.