Restyle action bar and tabs
Give the action bar and tabs a more identifiable and modern look. Not
fully styled yet (for example, the highlight color for the tabs) and
the action icons need changing.

A special style for a screen width of at least 480dp is applied to the
tabs so that they have white text when they are collapsed into the
action bar. Not the nicest solution, but there doesn't seem to be a
better way at the moment.