Add empty navigation drawer
The navigation drawer has recently become a standard design pattern and
it would certainly be very useful for this application, so add it to
the main activity. It is currently empty but will contain navigation
between the main areas of the application. The navigation drawer is
provided by version 13 of the support library.

A custom white drawer icon is used. It unfortunately clashes a little
with the application icon, but I have plans to change that anyway.