Add background drawable for card pattern
The card pattern is becoming quite popular with Android applications
and will probably be standardised soon. Add a drawable for the card
background which is actually two rounded rectangles layered on top
of each other. The result is a rounded rectangle with a 1px border
on all sides but the bottom which is 2px.

Change the sites list style to try it out, but this is likely going to
change a lot soon.