Improve sites list item layout
Change the list item layout for sites to include tags and a footer area
for account information. This will have much richer content than the
previous design. The plan is to show a personalised list of sites, such
that the user's favourite sites and tags show up at the top.