Add notification drawer
Add a notification drawer that slides in from the right-hand side when
the notification action button is pressed. It interacts nicely with the
navigation drawer, such that they are not both open at the same time. In
addition, the action bar menu icon should only animate for the
navigation drawer.