Move navigation drawer code to DrawerActivity
DrawerActivity should be extended by any activity that wishes to have
the navigation and notification drawers. For activities not at the top
level, call `setDrawerIndicatorEnabled(false)` to show the standard up
button in the action bar.

DrawerActivity will inflate a wrapper layout that adds the appropriate
drawer views. Calling `setContentView` from a subclass activity will
actually fill the content subview, so it can be used as usual.

When the an item in the navigation drawer is selected, the drawer will
close and only once it has fully closed will the appropriate intent be
fired. During this time, touch interaction is disabled, to prevent the
user from interrupting the drawer closing animation.