Select requested tab in `NetworkActivity.onCreate`
`onNewIntent` will only be fired when `NetworkActivity` is already at
the top of the task back stack. When navigation to `NetworkActivity`
from some other activity by selecting one of "What's hot?" and "Sites"
from the navigation drawer, the appropriate tab is not automatically
selected. Move the tab selection code to a separate function and call it
from both `onCreate` and `onNewIntent`.